fixed another bug
Former-commit-id: fd761f5583ce619cf13e70727ce4b9aa5bcfa0c4 [formerly fd761f5583ce619cf13e70727ce4b9aa5bcfa0c4 [formerly d2f73ef85ad925133af773d8d94513c9038b32b5]]
Former-commit-id: a7a86257ca6289df3e0cbc8c783bdd3c9b9e9619
Former-commit-id: ed9968cc814b078dd5c68009df9c854cf9ec6910

+++ arn/data/kinetics.py
@@ -23,6 +23,7 @@ def load_value_file(file_path):
 
 
 def my_video_loader(seq_path):
+
     frames = []
     # extract frames from the video
     if os.path.exists(seq_path):
@@ -39,7 +40,7 @@ def my_video_loader(seq_path):
     if len(frames) == 0:
         print(seq_path + " is busted")
         shutil.move(seq_path, "/media/sgrieggs/pageparsing/badvideos/"+seq_path.split("/")[-1])
-        frames = [np.zeros((360, 640, 3))]
+        frames = [np.zeros((360, 640, 3)).astype(np.uint8)]
     return frames
 
 def pil_loader(path):
